+++ R/OhdsiShinyModules.R
@@ -20,8 +20,8 @@
 #' 
 #' @description A selection of shiny modules for exploring standardized OHDSI results 
 #'
-#' @docType package
 #' @name OhdsiShinyModules
+#' @keywords internal
 #' @importFrom dplyr %>%
 #' @importFrom rlang .data
-
+"_PACKAGE"