+++ R/components-data-viewer.R
@@ -9,7 +9,7 @@
 #'
 #' @param id string
 #' @param downloadedFileName string, desired name of downloaded data file. can use the name from the module that is being used
-#'
+#' @param boxTitle the title added to the box 
 #' @return shiny module UI
 #'
 resultTableViewer <- function(