+++ R/estimation-sccs-diagnostics.R
@@ -133,7 +133,7 @@ estimationGetSccsDiagnostics <- function(
   
   # add summaryValue after outcome
   result <- result %>% 
-    dplyr::relocate(.data$summaryValue, .after = .data$outcome)
+    dplyr::relocate("summaryValue", .after = "outcome")
   
   return(
     result