+++ tests/testthat/test-estimation-cohort-method-propensityScoreDist.R
@@ -4,20 +4,20 @@ shiny::testServer(
   app = cohortMethodPropensityScoreDistServer, 
   args = list(
     selectedRow = shiny::reactiveVal(NULL), 
-    connectionHandler = connectionHandlerCm, 
-    resultDatabaseSettings = resultDatabaseSettingsCm
+    connectionHandler = connectionHandlerEstimation, 
+    resultDatabaseSettings = resultDatabaseSettingsEstimation
   ), 
   expr = {
     
     testthat::expect_true(is.null(psDistPlot()))
     
     ps <- getCohortMethodPs(
-      connectionHandler = connectionHandlerCm, 
-      resultDatabaseSettings = resultDatabaseSettingsCm,
+      connectionHandler = connectionHandlerEstimation, 
+      resultDatabaseSettings = resultDatabaseSettingsEstimation,
       targetId = 1, 
       comparatorId = 2, 
       analysisId = 2, 
-      databaseId = '1'
+      databaseId = 'eunomia'
     )
     
     testthat::expect_true('preferenceScore' %in% colnames(ps))
@@ -27,7 +27,7 @@ shiny::testServer(
     # make sure this runs if we pick the first row
     selectedRow(
       list(
-        databaseId = '1', 
+        databaseId = 'eunomia', 
         cdmSourceAbbreviation = 'Eunomia', 
         description  = 'madeup',
         target = 'test target',